PR feedback: typos & renaming
