+++ src/api/winnings/winnings.module.ts
@@ -6,7 +6,7 @@ import { WinningsRepository } from '../repository/winnings.repo';
 import { TaxFormRepository } from '../repository/taxForm.repo';
 import { PaymentMethodRepository } from '../repository/paymentMethod.repo';
 import { TopcoderModule } from 'src/shared/topcoder/topcoder.module';
-import { IdentityVerificationRepository } from '../repository/identiti-verification.repo';
+import { IdentityVerificationRepository } from '../repository/identity-verification.repo';
 
 @Module({
   imports: [TopcoderModule],
